+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/JsonRpcConfiguration.java
@@ -42,6 +42,7 @@ public class JsonRpcConfiguration {
   private String host;
   private List<String> corsAllowedDomains = Collections.emptyList();
   private List<String> rpcApis;
+  private List<String> noAuthRpcApis = Collections.emptyList();
   private List<String> hostsAllowlist = Arrays.asList("localhost", "127.0.0.1");
   private boolean authenticationEnabled = false;
   private String authenticationCredentialsFile;
@@ -114,6 +115,14 @@ public class JsonRpcConfiguration {
 
   public void setRpcApis(final List<String> rpcApis) {
     this.rpcApis = rpcApis;
+  }
+
+  public Collection<String> getNoAuthRpcApis() {
+    return this.noAuthRpcApis;
+  }
+
+  public void setNoAuthRpcApis(final List<String> rpcApis) {
+    this.noAuthRpcApis = rpcApis;
   }
 
   public void addRpcApi(final String rpcApi) {